requirements.txt
boto3==1.17.109
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
dnspython==1.16.0
drf-simplejwt==4.4.1
pandas==1.3.0
pymongo==3.11.4
requests==2.25.1
scikit-learn==0.24.2
sentry-sdk==1.1.0
